+++ config/default/DeliveryExecutionState.conf.php
@@ -1,0 +1,8 @@
+<?php
+use oat\taoProctoring\model\DeliveryExecutionStateService;
+/**
+ * Default authorization service
+ */
+return new DeliveryExecutionStateService([
+
+]);
