+++ helpers/TestCenterHelper.php
@@ -32,6 +32,7 @@ use oat\tao\helpers\UserHelper;
 use oat\taoQtiTest\models\TestSessionMetaData;
 use oat\taoProctoring\model\implementation\DeliveryService;
 use oat\taoProctoring\model\EligibilityService;
+use oat\taoProctoring\model\DeliveryExecutionStateService;
 
 /**
  * This temporary helpers is a temporary way to return data to the controller.
@@ -182,7 +183,9 @@ class TestCenterHelper
             }
         }
 
-        return DataTableHelper::paginate($filteredExecutions, $options, function($deliveryExecutions) use ($deliveryService) {
+        $deliveryExecutionStateService = ServiceManager::getServiceManager()->get(DeliveryExecutionStateService::SERVICE_ID);
+
+        return DataTableHelper::paginate($filteredExecutions, $options, function($deliveryExecutions) use ($deliveryExecutionStateService) {
             $reports = [];
 
             foreach($deliveryExecutions as $deliveryExecution) {
@@ -193,7 +196,7 @@ class TestCenterHelper
                 $userId = $deliveryExecution->getUserIdentifier();
                 $user = UserHelper::getUser($userId);
 
-                $state = $deliveryService->getProctoringState($deliveryExecution->getUri());
+                $state = $deliveryExecutionStateService->getProctoringState($deliveryExecution->getUri());
                 $proctor = null;
                 if (!empty($state['authorized_by'])) {
                     $proctor = UserHelper::getUser($state['authorized_by']);
@@ -205,7 +208,7 @@ class TestCenterHelper
                     'delivery' => $deliveryExecution->getDelivery()->getLabel(),
                     'testtaker' => $user ? UserHelper::getUserName($user, true) : '',
                     'proctor' => $proctor ? UserHelper::getUserName($proctor, true) : '',
-                    'status' => $deliveryService->getState($deliveryExecution),
+                    'status' => $deliveryExecutionStateService->getState($deliveryExecution),
                     'start' => $startTime ? DateHelper::displayeDate($startTime) : '',
                     'end' => $finishTime ? DateHelper::displayeDate($finishTime) : '',
                     'pause' => $procActions['pause'],
@@ -224,8 +227,8 @@ class TestCenterHelper
      */
     protected static function getProctorActions($deliveryExecution)
     {
-        $ds = ServiceManager::getServiceManager()->get(DeliveryService::CONFIG_ID);
-        $session = $ds->getTestSession($deliveryExecution);
+        $deliveryExecutionStateService = ServiceManager::getServiceManager()->get(DeliveryExecutionStateService::SERVICE_ID);
+        $session = $deliveryExecutionStateService->getTestSession($deliveryExecution);
         
         $actions = array(
             'pause' => 0,
